# Conflicts:
#	core/cpu.c

+++ core/apb.h
@@ -22,8 +22,8 @@ extern apb_map_entry_t apb_map[0x10];
 
 void apb_set_map(int entry, eZ80portrange_t* range);
 
-uint8_t port_read_byte(const uint32_t addr);
-void port_write_byte(const uint32_t addr, const uint8_t value);
+uint8_t port_read_byte(const uint16_t addr);
+void port_write_byte(const uint16_t addr, const uint8_t value);
 
 #ifdef __cplusplus
 }
